Updated Chart Percentage & Decimal on QTY

+++ backend/src/models/transaction_items.js
@@ -22,7 +22,7 @@ module.exports = (sequelize, DataTypes) => {
         product: DataTypes.INTEGER,
         price: DataTypes.DECIMAL(16,4),
         pricing_type: DataTypes.STRING,
-        quantity: DataTypes.INTEGER,
+        quantity: DataTypes.DECIMAL(16,4),
         created_by: DataTypes.INTEGER
     }, {
         sequelize,
